+++ importer/src/main/java/org/mskcc/cbio/importer/config/internal/GDataImpl.java
@@ -327,7 +327,11 @@ class GDataImpl implements Config {
 					// a delimited list of datatypes have been requested
 					toReturn = new ArrayList<DatatypeMetadata>();
 					for (String datatype : datatypesIndicator.split(DatatypeMetadata.DATATYPES_DELIMITER)) {
-						toReturn.add(getDatatypeMetadata(datatype).iterator().next());
+						DatatypeMetadata datatypeMetadata = getDatatypeMetadata(datatype).iterator().next();
+						toReturn.add(datatypeMetadata);
+						if (LOG.isInfoEnabled()) {
+							LOG.info("Selecting data type"+datatypeMetadata.getDatatype());
+                                                }
 					}
 				}
 				break;