+++ portal/src/main/resources/content/news.markdown
@@ -1,6 +1,7 @@
-# March 7, 2013
+# March 28, 2013
 
-* All data updated to the January Firehose run (January 16, 2012).
+* All data TCGA updated to the January Firehose run (January 16, 2012).
+* Data from a new bladder cancer study from MSKCC has been added (97 samples, Iyer et al., JCO in press).
 
 # February 16, 2013